# Speed up package search