--- a/services/DevHome.Services.WindowsPackageManager/Services/Operations/WinGetGetPackageOperation.cs
+++ b/services/DevHome.Services.WindowsPackageManager/Services/Operations/WinGetGetPackageOperation.cs
@@ -65,13 +65,8 @@ internal sealed class WinGetGetPackageOperation : IWinGetGetPackageOperation
                     // Get packages from the catalog
                     var catalog = await _protocolParser.ResolveCatalogAsync(firstParsedUri);
                     var packagesOutOfProc = await _packageFinder.GetPackagesAsync(catalog, packageIds);
-                    var packagesInProc = packagesOutOfProc
-                        .AsParallel()
-                        .AsOrdered()
-                        .Select(p => new WinGetPackage(_logger, p, _packageFinder.IsElevationRequired(p)))
-                        .ToList<IWinGetPackage>();
-                    packagesInProc
-                        .ForEach(p => _packageCache.TryAddPackage(_protocolParser.CreatePackageUri(p), p));
+                    var packagesInProc = WinGetPackage.FromOutOfProc(_logger, _packageFinder, packagesOutOfProc);
+                    packagesInProc.ForEach(p => _packageCache.TryAddPackage(_protocolParser.CreatePackageUri(p), p));
                     return packagesInProc;
                 }));
             }