--- a/services/DevHome.Services.WindowsPackageManager/Services/Operations/WinGetSearchOperation.cs
+++ b/services/DevHome.Services.WindowsPackageManager/Services/Operations/WinGetSearchOperation.cs
@@ -40,11 +40,7 @@ internal sealed class WinGetSearchOperation : IWinGetSearchOperation
         {
             var searchCatalog = await _catalogConnector.GetCustomSearchCatalogAsync();
             var results = await _packageFinder.SearchAsync(searchCatalog, query, limit);
-            return results
-                .AsParallel()
-                .AsOrdered()
-                .Select(p => new WinGetPackage(_logger, p, _packageFinder.IsElevationRequired(p)))
-                .ToList<IWinGetPackage>();
+            return WinGetPackage.FromOutOfProc(_logger, _packageFinder, results);
         });
     }
 }