## Summary of the pull request

Code changes in this PR allows copying packages from out-of-proc to in-proc in parallel offering x1.44 speedup on average per search query. Those values are based on 28 data points executed before and after the change.

⚠️ Those numbers may be different on different machines and depends on hardware and software variables.

  | Before (ms) | After (ms) | Speedup
-- | -- | -- | --
SUM | 46,613 | 32,272 | 1.444379
AVG | 1,726 | 1,195 | 1.444379

## References and relevant issues

## Detailed description of the pull request / Additional comments

## Validation steps performed

## PR checklist
- [ ] Closes #3655
- [ ] Tests added/passed
- [ ] Documentation updated